Refactor response data transformation to utility class
This commit introduces a new utility class, `ResponseTransformer`, which is responsible for transforming response data keys to snake_case and converting Date instances to ISO strings. This class has been used to replace the in-place data transformation in `TransformResponseInterceptor`. Also, it has been integrated into `AllExceptionsFilter` to ensure that exception responses are also transformed. Changes include modifications to the `TransformResponseInterceptor` to leverage the new utility class, the addition of the `ResponseTransformer` utility class, and adjustments to tests and other parts of the codebase to match the changes in the response structure. All constraint keys and validation error keys have been updated to snake_case as well.